+++ src/Resources/contao/dca/tl_survey.php
@@ -18,7 +18,6 @@ use Contao\Backend;
 use Contao\BackendUser;
 use Contao\Database;
 use Contao\Input;
-use Doctrine\DBAL\Platforms\AbstractMySQLPlatform;
 
 /*
  * Table tl_survey
@@ -103,8 +102,8 @@ $GLOBALS['TL_DCA']['tl_survey'] = [
         '__selector__' => ['access', 'limit_groups', 'useResultCategories'],
         'default' => '{title_legend},title,author,description,language;{activation_legend},online_start,online_end',
         'anon' => '{title_legend},title,author,description,language;{activation_legend},online_start,online_end;{access_legend},access,usecookie;{texts_legend},introduction,finalsubmission;{head_legend},show_title,show_cancel;{sendconfirmationmail_legend:hide},sendConfirmationMail,sendConfirmationMailAlternate;{misc_legend},allowback,immediate_start,jumpto,useResultCategories',
-        'anoncode' => '{title_legend},title,author,description,language;{activation_legend},online_start,online_end;{access_legend},access,usecookie;{texts_legend},introduction,finalsubmission;{head_legend},show_title,show_cancel;{sendconfirmationmail_legend:hide},sendConfirmationMail,sendConfirmationMailAlternate;{misc_legend},allowback,immediate_start,jumpto,useResultCategories',
-        'nonanoncode' => '{title_legend},title,author,description,language;{activation_legend},online_start,online_end;{access_legend},access,usecookie,limit_groups;{texts_legend},introduction,finalsubmission;{head_legend},show_title,show_cancel;{sendconfirmationmail_legend:hide},sendConfirmationMail,sendConfirmationMailAlternate;{misc_legend},allowback,immediate_start,jumpto,useResultCategories',
+        'anoncode' => '{title_legend},title,author,description,language;{activation_legend},online_start,online_end;{access_legend},access,usecookie;{texts_legend},introduction,finalsubmission;{head_legend},show_title,show_cancel;{sendconfirmationmail_legend:hide},sendConfirmationMail,sendConfirmationMailAlternate;{misc_legend},allowback,immediate_start,allow_autostart,jumpto,useResultCategories',
+        'nonanoncode' => '{title_legend},title,author,description,language;{activation_legend},online_start,online_end;{access_legend},access,usecookie,limit_groups;{texts_legend},introduction,finalsubmission;{head_legend},show_title,show_cancel;{sendconfirmationmail_legend:hide},sendConfirmationMail,sendConfirmationMailAlternate;{misc_legend},allowback,immediate_start,allow_autostart,jumpto,useResultCategories',
     ],
 
     // Palettes
@@ -239,19 +238,24 @@ $GLOBALS['TL_DCA']['tl_survey'] = [
             'filter' => true,
             'exclude' => true,
             'inputType' => 'checkbox',
-            'eval' => ['tl_class' => 'w50'],
+            'eval' => ['tl_class' => 'w33'],
             'sql' => "char(1) NOT NULL default ''",
         ],
         'immediate_start' => [
-            'label' => &$GLOBALS['TL_LANG']['tl_survey']['immediate_start'],
             'filter' => true,
             'exclude' => true,
             'inputType' => 'checkbox',
-            'eval' => ['tl_class' => 'w50'],
+            'eval' => ['tl_class' => 'w33'],
+            'sql' => "char(1) NOT NULL default ''",
+        ],
+        'allow_autostart' => [
+            'filter' => true,
+            'exclude' => true,
+            'inputType' => 'checkbox',
+            'eval' => ['tl_class' => 'w33'],
             'sql' => "char(1) NOT NULL default ''",
         ],
         'jumpto' => [
-            'label' => &$GLOBALS['TL_LANG']['tl_survey']['jumpto'],
             'exclude' => true,
             'inputType' => 'pageTree',
             'explanation' => 'jumpTo',
@@ -259,7 +263,6 @@ $GLOBALS['TL_DCA']['tl_survey'] = [
             'sql' => "int(10) unsigned NOT NULL default '0'",
         ],
         'surveyPage' => [
-            'label' => &$GLOBALS['TL_LANG']['tl_survey']['surveyPage'],
             'inputType' => 'pageTree',
             'eval' => ['mandatory' => false, 'fieldType' => 'radio'],
         ],
@@ -285,7 +288,6 @@ $GLOBALS['TL_DCA']['tl_survey'] = [
             ],
             'sql' => [
                 'type' => 'blob',
-                'length' => AbstractMySQLPlatform::LENGTH_LIMIT_BLOB,
                 'notnull' => false,
             ],
         ],